requirements.txt
alembic==1.9.4
anyio==3.6.2
bcrypt==4.0.1
boto3==1.26.123
botocore==1.29.123
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
docutils==0.14
email-validator==2.0.0.post2
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
jmespath==1.0.1
lockfile==0.11.0
Mako==1.2.4
MarkupSafe==2.1.2
pbr==0.11.1
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-daemon==2.2.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
s3transfer==0.6.0
simplejson==3.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
zipp==3.14.0
